Moved metric hooks to DataLayer domain
