+++ js/src/domains/interventions/components/InterventionFormWrapper.tsx
@@ -4,8 +4,8 @@ import { Stack, Typography } from '@mui/material';
 import { LoadingSpinner, useSafeIntl } from 'bluesquare-components';
 import { CardStyled } from '../../../components/CardStyled';
 import { ExtendedFormikProvider } from '../../../hooks/useGetExtendedFormikContext';
+import { useGetMetricTypes } from '../../dataLayers/hooks/useGetMetrics';
 import { MESSAGES } from '../../messages';
-import { useGetMetricTypes } from '../../planning/hooks/useGetMetrics';
 import { InterventionDetails } from '../../planning/types/interventions';
 import { useGetBudgetSettings } from '../hooks/useGetBudgetSettings';
 import { useGetInterventionCostBreakdownLineCategories } from '../hooks/useGetInterventionCostBreakdownLineCategories';
