+++ js/src/domains/planning/components/interventionPlanMap/InterventionPlanMap.tsx
@@ -4,10 +4,10 @@ import { OrgUnit } from 'Iaso/domains/orgUnits/types/orgUnit';
 import { SxStyles } from 'Iaso/types/general';
 import { LayerSelect } from '../../../../components/LayerSelect';
 import { Map as SNTMap } from '../../../../components/Map';
+import { useGetMetricValues } from '../../../dataLayers/hooks/useGetMetrics';
 import { MetricType } from '../../../dataLayers/types/metrics';
 import { MESSAGES } from '../../../messages';
 import { usePlanningContext } from '../../contexts/PlanningContext';
-import { useGetMetricValues } from '../../hooks/useGetMetrics';
 import {
     getMapStyleForOrgUnit,
     useGetOrgUnitMetric,
